+++ demo/CypressLayout.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import * as L from '../leda';
 import {
-  AutoComplete, Button, CheckBox, MaskedInput, VStepper, ButtonGroup, Input, DatePicker, DateRange, DropDownSelect, 
+  AutoComplete, Button, CheckBox, MaskedInput, VStepper, ButtonGroup, Input, DatePicker, DateRange, DropDownSelect, DropZone,
 } from './cypress';
 import { Styles } from './components/Styles';
 
@@ -40,6 +40,9 @@ export const CypressLayout = () => (
         </Route>
         <Route path="/cypress/vstepper">
           <VStepper />
+        </Route>
+        <Route path="/cypress/dropzone">
+          <DropZone />
         </Route>
       </Switch>
     </L.Div>