+++ demo/cypress/index.ts
@@ -9,4 +9,5 @@ export { VStepper } from './VStepper';
 export { DatePicker } from './DatePicker';
 export { DateRange } from './DateRange';
 export { DropDownSelect } from './DropDownSelect';
+export { DropZone } from './DropZone';