+++ leda/components/DropZone/DropZone.test.tsx
@@ -1,44 +1,26 @@
-// @ts-nocheck
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 import React from 'react';
 import { mount } from 'enzyme';
+import {
+  render, fireEvent, act, wait,
+} from '@testing-library/react';
 import toJson from 'enzyme-to-json';
-import { act } from 'react-dom/test-utils';
 import { DropZone } from './index';
-import { Li } from '../Li';
+import { FileErrorCodes } from './types';
 
-let files;
-let firstFile;
-let secondFile;
-let container;
+const files = [
+  new File(['file 1'], 'file_1.txt', { type: 'plain/text' }),
+  new File(['file 2'], 'file_2.pdf', { type: 'application/pdf' }),
+];
 
-beforeEach(() => {
-  files = [
-    {
-      name: 'file1.pdf',
-      size: 19000000,
-      type: 'application/pdf',
-    },
-    {
-      name: 'cats.gif',
-      size: 21000000,
-      type: 'image/gif',
-    },
-  ];
+const externalFile = { name: 'file_3.txt', link: 'http://example.com/file.txt' };
 
-  [firstFile, secondFile] = files;
+const rejectedFile = new File(['rejected file'], 'rejected_file.txt', { type: 'plain/text' });
+Object.assign(rejectedFile, { errorCode: FileErrorCodes.FileIsTooBig });
 
-  window.URL.createObjectURL = jest.fn();
+const mockObjectURL = 'blob://mockObjectURL';
 
-  container = document.createElement('div');
-
-  document.body.appendChild(container);
-});
-
-afterEach(() => {
-  document.body.removeChild(container);
-
-  container = null;
+beforeAll(() => {
+  window.URL.createObjectURL = jest.fn(() => mockObjectURL);
 });
 
 describe('DropZone SNAPSHOTS', () => {
@@ -47,10 +29,133 @@ describe('DropZone SNAPSHOTS', () => {
 
     expect(toJson(wrapper)).toMatchSnapshot();
   });
+
+  it('should render file list', () => {
+    const { container } = render(
+      <DropZone
+        value={{
+          acceptedFiles: [
+            ...files,
+            externalFile,
+          ],
+          rejectedFiles: [rejectedFile],
+        }}
+      />,
+    );
+
+    expect(container).toMatchSnapshot();
+  });
 });
 
 describe('DropZone HANDLERS', () => {
+  it('should call onClick handler', () => {
+    const handleClick = jest.fn();
 
+    const { getByText } = render(<DropZone onClick={handleClick} />);
+
+    fireEvent.click(getByText('Выбрать...'));
+
+    expect(handleClick).toHaveBeenCalledTimes(1);
+  });
+
+  it('should call onChange handler on drop files into the dropZone', async () => {
+    const handleChange = jest.fn();
+
+    const eventMatcher = expect.objectContaining({
+      component: expect.objectContaining({
+        dropped: {
+          acceptedFiles: [
+            expect.objectContaining(files[0]),
+            expect.objectContaining(files[1]),
+          ],
+          rejectedFiles: [],
+        },
+        value: {
+          acceptedFiles: [
+            expect.objectContaining(files[0]),
+            expect.objectContaining(files[1]),
+          ],
+          rejectedFiles: [],
+        },
+      }),
+    });
+
+    const ui = <DropZone onChange={handleChange} />;
+
+    const { rerender, container, queryByText } = render(ui);
+
+    const dropZoneContent = container.querySelector('div.dropzone-content');
+
+    expect(dropZoneContent).toBeInTheDocument();
+
+    const data = {
+      dataTransfer: {
+        files,
+        items: files.map((file) => ({
+          kind: 'file',
+          type: file.type,
+          getAsFile: () => file,
+        })),
+        types: ['Files'],
+      },
+    };
+
+    const event = new Event('drop', { bubbles: true });
+    Object.assign(event, data);
+
+    fireEvent(dropZoneContent as Element, event);
+
+    await act(() => wait(() => rerender(ui)));
+
+    expect(queryByText('file_1.txt')).toBeInTheDocument();
+    expect(queryByText('file_2.pdf')).toBeInTheDocument();
+
+    expect(handleChange).toHaveBeenCalledTimes(1);
+    expect(handleChange).toHaveBeenLastCalledWith(eventMatcher);
+  });
+
+  /*
+   TODO fix the bug with incorrect "dropped" value in the change event triggered by removing file
+   Now: "dropped" equals "value" before remove
+   Expected: "dropped" empty
+   */
+  // eslint-disable-next-line jest/no-disabled-tests
+  it.skip('should call onChange handler on remove file', () => {
+    const handleChange = jest.fn();
+
+    const eventMatcher = expect.objectContaining({
+      component: expect.objectContaining({
+        dropped: {
+          acceptedFiles: [],
+          rejectedFiles: [],
+        },
+        value: {
+          acceptedFiles: [],
+          rejectedFiles: [],
+        },
+        removedFile: [externalFile],
+      }),
+    });
+
+    const { container } = render(
+      <DropZone
+        onChange={handleChange}
+        value={{
+          acceptedFiles: [externalFile],
+          rejectedFiles: [],
+        }}
+      />,
+    );
+
+    const removeFileIcon = container.querySelector('i.dropzone-delete-icon');
+
+    expect(removeFileIcon).toBeInTheDocument();
+
+    fireEvent.click(removeFileIcon as Element);
+
+    expect(handleChange).toHaveBeenCalledTimes(1);
+    expect(handleChange).toHaveBeenLastCalledWith(eventMatcher);
+  });
 });
 
 describe('DropZone ATTRIBUTES', () => {
@@ -83,8 +188,8 @@ describe('DropZone ATTRIBUTES', () => {
 
     wrapper.setProps({ value: { acceptedFiles: files, rejectedFiles: [] } });
 
-    expect(div.innerHTML.includes('file1.pdf')).toBeTruthy();
-    expect(div.innerHTML.includes('cats.gif')).toBeTruthy();
+    expect(div.innerHTML.includes('file_1.txt')).toBeTruthy();
+    expect(div.innerHTML.includes('file_2.pdf')).toBeTruthy();
   });
 
   it('should render descriptionText', () => {