+++ leda/components/DropZone/constants.ts
@@ -1,0 +1,4 @@
+export const EMPTY_DROP_ZONE_FILES = {
+  acceptedFiles: [],
+  rejectedFiles: [],
+};