+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.31.0] - 2020-09-28
 
+### Fixed
+- DropZone: Fixed downloading external and attached files
+
+
 ## [0.30.0] - 2020-09-14
 
 ### Added